storage: Move resize related code to its own file
This will be used also for partitions in a future commit.
